requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
aiosqlite==0.19.0
supabase==2.9.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI and ML
httpx==0.27.2
google-generativeai==0.3.2
replicate==0.22.0

# Image Processing
Pillow==11.0.0
opencv-python-headless==4.10.0.84

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dateutil==2.9.0

# SSH remote management (for run_camera.py - dev machine only)
paramiko==3.4.0

# Hardware (for Raspberry Pi - optional)
gpiozero==2.0.1; platform_machine == "armv7l" or platform_machine == "aarch64"
picamera2==0.3.16; platform_machine == "armv7l" or platform_machine == "aarch64"

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
